+++ networking/p2p/src/main/java/tech/pegasys/teku/networking/p2p/discovery/discv5/SecretKeyParser.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright ConsenSys Software Inc., 2022
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.networking.p2p.discovery.discv5;
+
+import java.math.BigInteger;
+import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.crypto.SECP256K1.SecretKey;
+
+public class SecretKeyParser {
+
+  public static SecretKey fromLibP2pPrivKey(final Bytes key) {
+    return SecretKey.fromInteger(new BigInteger(key.toArrayUnsafe()));
+  }
+}